move add delim command to management command dir for s3 improvment
