+++ addons/s3/utils.py
@@ -10,9 +10,6 @@ from boto.s3.connection import OrdinaryCallingFormat
 from framework.exceptions import HTTPError
 from addons.base.exceptions import InvalidAuthError, InvalidFolderError
 from addons.s3.settings import BUCKET_LOCATIONS
-from django.apps import apps
-from django.db.models import F, Value
-from django.db.models.functions import Concat, Replace
 
 logger = logging.getLogger(__name__)
 
@@ -156,47 +153,3 @@ def get_bucket_prefixes(access_key, secret_key, prefix, bucket_name):
             )
 
     return folders
-
-
-def update_folder_names():
-    NodeSettings = apps.get_model('addons_s3', 'NodeSettings')
-
-    # Update folder_id for all records
-    NodeSettings.objects.exclude(
-        folder_name__contains=':/'
-    ).update(
-        folder_id=Concat(F('folder_id'), Value(':/'))
-    )
-
-    # Update folder_name for records containing '('
-    NodeSettings.objects.filter(
-        folder_name__contains=' ('
-    ).exclude(
-        folder_name__contains=':/'
-    ).update(
-        folder_name=Replace(F('folder_name'), Value(' ('), Value(':/ ('))
-    )
-    NodeSettings.objects.exclude(
-        folder_name__contains=':/'
-    ).exclude(
-        folder_name__contains=' ('
-    ).update(
-        folder_name=Concat(F('folder_name'), Value(':/'))
-    )
-    logger.info('Update Folder Names/IDs complete')
-
-
-def reverse_update_folder_names():
-    NodeSettings = apps.get_model('addons_s3', 'NodeSettings')
-
-    # Reverse update folder_id for all records
-    NodeSettings.objects.update(folder_id=Replace(F('folder_id'), Value(':/'), Value('')))
-
-    # Reverse update folder_name for records containing ':/ ('
-    NodeSettings.objects.filter(folder_name__contains=':/ (').update(
-        folder_name=Replace(F('folder_name'), Value(':/ ('), Value(' ('))
-    )
-    NodeSettings.objects.filter(folder_name__contains=':/').update(
-        folder_name=Replace(F('folder_name'), Value(':/'), Value(''))
-    )
-    logger.info('Reverse Update Folder Names/IDs complete')
